v1.6.1 support save file

+++ app/src/main/java/com/study/qrscanner/QRGenerating/BizCardGnrActivity.java
@@ -5,9 +5,9 @@ import android.graphics.Bitmap;
 import android.graphics.Point;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
+import android.os.Bundle;
 import android.os.Environment;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.util.Log;
 import android.view.Display;
 import android.view.Menu;
@@ -41,10 +41,10 @@ public class BizCardGnrActivity extends AppCompatActivity {
         final String qrInputText = QRData.getString("gn");
 
 
-        Bitmap bitmap=createBitmap(qrInputText);
+        Bitmap bitmap = createBitmap(qrInputText);
 
         ImageView myImage = (ImageView) findViewById(R.id.imageView1);
-        myImage.setImageBitmap(bitmap );
+        myImage.setImageBitmap(bitmap);
 
 
         btnstore.setOnClickListener(new View.OnClickListener() {
@@ -56,14 +56,12 @@ public class BizCardGnrActivity extends AppCompatActivity {
                 final String qrInputText = QRData.getString("gn");
 
 
-                Bitmap bitmap=createBitmap(qrInputText);
+                Bitmap bitmap = createBitmap(qrInputText);
                 saveImageToExternalStorage(bitmap);
 
-                Intent i=new Intent(BizCardGnrActivity.this, ScannerActivity.class);
+                Intent i = new Intent(BizCardGnrActivity.this, ScannerActivity.class);
                 startActivity(i);
                 Toast.makeText(BizCardGnrActivity.this, "QR code stored in gallery", Toast.LENGTH_LONG).show();
-
-
 
 
             }
@@ -73,7 +71,7 @@ public class BizCardGnrActivity extends AppCompatActivity {
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.share,menu);
+        getMenuInflater().inflate(R.menu.share, menu);
 
 
         return true;
@@ -83,12 +81,11 @@ public class BizCardGnrActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
 
 
-
-        switch (item.getItemId()){
+        switch (item.getItemId()) {
             case R.id.share:
                 Bundle QRData = getIntent().getExtras();//from QRGenerator
                 final String qrInputText = QRData.getString("gn");
-                Bitmap bitmap=createBitmap(qrInputText);
+                Bitmap bitmap = createBitmap(qrInputText);
                 shareIt(bitmap);
 
                 return true;
@@ -97,10 +94,10 @@ public class BizCardGnrActivity extends AppCompatActivity {
                 return super.onOptionsItemSelected(item);
         }
     }
-    public void shareIt(Bitmap result)
-    {
+
+    public void shareIt(Bitmap result) {
         try {
-            File file = new File(this.getExternalCacheDir(),"logicchip.png");
+            File file = new File(this.getExternalCacheDir(), "logicchip.png");
             FileOutputStream fOut = new FileOutputStream(file);
             result.compress(Bitmap.CompressFormat.PNG, 100, fOut);
             fOut.flush();
@@ -118,8 +115,6 @@ public class BizCardGnrActivity extends AppCompatActivity {
     }
 
 
-
-
     private void saveImageToExternalStorage(Bitmap finalBitmap) {
         String root = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES).toString();
         File myDir = new File(root + "/saved_images");
@@ -129,31 +124,29 @@ public class BizCardGnrActivity extends AppCompatActivity {
         n = generator.nextInt(n);
         String fname = "Image-" + n + ".jpg";
         File file = new File(myDir, fname);
-        if (file.exists())
-            file.delete();
+        if (file.exists()) file.delete();
         try {
             FileOutputStream out = new FileOutputStream(file);
             finalBitmap.compress(Bitmap.CompressFormat.JPEG, 90, out);
             out.flush();
             out.close();
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
         }
 
 
         // Tell the media scanner about the new file so that it is
         // immediately available to the user.
-        MediaScannerConnection.scanFile(this, new String[] { file.toString() }, null,
-                new MediaScannerConnection.OnScanCompletedListener() {
-                    public void onScanCompleted(String path, Uri uri) {
-                        Log.i("ExternalStorage", "Scanned " + path + ":");
-                        Log.i("ExternalStorage", "-> uri=" + uri);
-                    }
-                });
+        MediaScannerConnection.scanFile(this, new String[]{file.toString()}, null, new MediaScannerConnection.OnScanCompletedListener() {
+            public void onScanCompleted(String path, Uri uri) {
+                Log.i("ExternalStorage", "Scanned " + path + ":");
+                Log.i("ExternalStorage", "-> uri=" + uri);
+            }
+        });
 
     }
-    public Bitmap createBitmap (String qrInputText) {
+
+    public Bitmap createBitmap(String qrInputText) {
 
         //Find screen size
         WindowManager manager = (WindowManager) getSystemService(WINDOW_SERVICE);
@@ -166,11 +159,7 @@ public class BizCardGnrActivity extends AppCompatActivity {
         smallerDimension = smallerDimension * 3 / 4;
 
         //Encode with a QR Code image
-        QRCodeEncoder qrCodeEncoder = new QRCodeEncoder(qrInputText,
-                null,
-                Contents.Type.Biz_Card,
-                BarcodeFormat.QR_CODE.toString(),
-                smallerDimension);
+        QRCodeEncoder qrCodeEncoder = new QRCodeEncoder(qrInputText, null, Contents.Type.Biz_Card, BarcodeFormat.QR_CODE.toString(), smallerDimension);
         Bitmap bitmap_ = null;
         try {
             bitmap_ = qrCodeEncoder.encodeAsBitmap();
